Move PositiveReal into hoomd_utility::valid
