+++ hoomd-vector/Cargo.toml
@@ -17,6 +17,7 @@ workspace = true
 [dependencies]
 thiserror = "2.0.12"
 rand = "0.9.0"
+hoomd-utility = { path = "../hoomd-utility" }
 
 [dev-dependencies]
 approx = "0.5.1"
